Refactor: Reorganized files into meaningful folders
